+++ CHANGES.md
@@ -38,6 +38,7 @@ Changes in 12.5
 - JS remove inline code from RegistrationAdminContactEnable() in RegistrationAdmin.fnc.php & RegistrationAdminContactEnable.js
 - JS remove inline code from RegistrationSiblingUseContactsAddress() in Registration.fnc.php & RegistrationSiblingUseContactsAddress.js
 - JS remove inline code from program in MassCreateAssignments.php & MassCreateAssignments.js
+- JS remove inline code from StudentAssignmentSubmissionOutput() in StudentAssignments.fnc.php & StudentAssignmentSubmissionOutput.js
 
 Changes in 12.4.3
 -----------------